Support array loader
Symfony allows loading a translation resource from a PHP array using
`Symfony\Component\Translation\Loader\ArrayLoader` but Kdyby\Translation
does not recognize that loader.

This patch adds the loader to the list of loader. Additionally,
because the resource value is used as an array key in the Tracy panel,
the value had to be serialized.